Revert "feat: PKCE"
