+++ .eslintrc
@@ -59,7 +59,8 @@
         "import/newline-after-import": "error",
         "import/no-duplicates": "error",
         "@typescript-eslint/no-floating-promises": 2,
-        "@typescript-eslint/no-unused-vars": 2
+        "@typescript-eslint/no-unused-vars": 2,
+        "typesafe/no-throw-sync-func": "error"
       }
     },
     { "files": ["*.spec.ts"], "extends": ["plugin:jest/recommended"] },
