+++ test/SgidClient.spec.ts
@@ -1,20 +1,15 @@
 import { readFileSync } from 'fs'
 
 import SgidClient from '../src'
-import {
-  generateCodeChallenge,
-  generateCodeVerifier,
-  generatePkcePair,
-} from '../src/generators'
 
 import {
   MOCK_ACCESS_TOKEN,
+  MOCK_API_VERSION,
   MOCK_AUTH_CODE,
   MOCK_AUTH_ENDPOINT,
   MOCK_CLIENT_ID,
   MOCK_CLIENT_PRIVATE_KEY,
   MOCK_CLIENT_SECRET,
-  MOCK_CODE_VERIFIER,
   MOCK_HOSTNAME,
   MOCK_REDIRECT_URI,
   MOCK_SUB,
@@ -28,15 +23,9 @@ import {
   userInfoHandlerNoData,
   userInfoHandlerNoKey,
 } from './mocks/handlers'
-import { codeVerifierAndChallengePattern } from './mocks/helpers'
 import { server } from './mocks/server'
 
-/**
- * Constants are redefined instead of being imported from "/src" so as to ensure any changes to either (but not both) would cause tests to fail.
- * This is to ensure that any changes we make to these are deliberate and not accidental.
- */
-const DEFAULT_SCOPE = 'myinfo.name openid'
-const DEFAULT_SGID_CODE_CHALLENGE_METHOD = 'S256'
+const DEFAULT_SCOPE = 'myinfo.nric_number openid'
 const DEFAULT_RESPONSE_TYPE = 'code'
 
 describe('SgidClient', () => {
@@ -49,6 +38,7 @@ describe('SgidClient', () => {
       privateKey: MOCK_CLIENT_PRIVATE_KEY,
       redirectUri: MOCK_REDIRECT_URI,
       hostname: MOCK_HOSTNAME,
+      apiVersion: MOCK_API_VERSION,
     })
   })
 
@@ -68,6 +58,7 @@ describe('SgidClient', () => {
         privateKey: pkcs8Key,
         redirectUri: MOCK_REDIRECT_URI,
         hostname: MOCK_HOSTNAME,
+        apiVersion: MOCK_API_VERSION,
       })
 
       expect(pkcs8Client).toBeDefined()
@@ -75,14 +66,10 @@ describe('SgidClient', () => {
   })
 
   describe('authorizationUrl', () => {
-    it('should generate authorisation URL correctly when state and codeChallenge are provided', () => {
+    it('should generate authorisation URL correctly when only state is provided', () => {
       const mockState = 'mockState'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(mockState)
 
       const actual = new URL(url)
       // Count number of search params
@@ -99,24 +86,15 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(8)
+      // Client ID, scope, response_type, redirect_uri, nonce, state
+      expect(actualNumSearchParams).toBe(6)
     })
 
-    it('should generate authorisation URL correctly when state, codeChallenge, and scope is provided as a string', () => {
+    it('should generate authorisation URL correctly when scope is provided as a string', () => {
       const mockState = 'mockState'
       const mockScope = 'mockScope'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        scope: mockScope,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(mockState, mockScope)
 
       const actual = new URL(url)
       // Count number of search params
@@ -134,24 +112,15 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(8)
+      // Client ID, scope, response_type, redirect_uri, nonce, state
+      expect(actualNumSearchParams).toBe(6)
     })
 
-    it('should generate authorisation URL correctly when state, codeChallenge, and scope is provided as a string array', () => {
+    it('should generate authorisation URL correctly when scope is provided as a string array', () => {
       const mockState = 'mockState'
       const mockScopes = ['mockScope1', 'mockScope2', 'mockScope3']
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        scope: mockScopes,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(mockState, mockScopes)
 
       const actual = new URL(url)
       // Count number of search params
@@ -169,24 +138,19 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(MOCK_REDIRECT_URI)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(8)
+      // Client ID, scope, response_type, redirect_uri, nonce, state
+      expect(actualNumSearchParams).toBe(6)
     })
 
-    it('should generate authorisation URL correctly when state, codeChallenge, and nonce is specified', () => {
+    it('should generate authorisation URL correctly when nonce is specified', () => {
       const mockState = 'mockState'
       const mockNonce = 'mockNonce'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        nonce: mockNonce,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(
+        mockState,
+        undefined,
+        mockNonce,
+      )
 
       const actual = new URL(url)
       // Count number of search params
@@ -205,23 +169,14 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('nonce')).toBe(mockNonce)
       expect(nonce).toBe(mockNonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(8)
+      // Client ID, scope, response_type, redirect_uri, nonce, state
+      expect(actualNumSearchParams).toBe(6)
     })
 
-    it('should generate authorisation URL correctly when state and codeChallenge is provided and nonce is null', () => {
+    it('should generate authorisation URL correctly when nonce is null', () => {
       const mockState = 'mockState'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        nonce: null,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(mockState, undefined, null)
 
       const actual = new URL(url)
       // Count number of search params
@@ -240,24 +195,20 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('nonce')).toBeNull()
       expect(nonce).toBeUndefined()
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(7)
+      // Client ID, scope, response_type, redirect_uri, state
+      expect(actualNumSearchParams).toBe(5)
     })
 
-    it('should generate authorisation URL correctly when state, codeChallenge, and redirectUri is provided', () => {
+    it('should generate authorisation URL correctly when redirectUri is provided', () => {
       const mockState = 'mockState'
       const mockRedirectUri = 'https://mockRedirectUri.com'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
-      const { url, nonce } = client.authorizationUrl({
-        state: mockState,
-        redirectUri: mockRedirectUri,
-        codeChallenge: mockCodeChallenge,
-      })
+      const { url, nonce } = client.authorizationUrl(
+        mockState,
+        undefined,
+        undefined,
+        mockRedirectUri,
+      )
 
       const actual = new URL(url)
       // Count number of search params
@@ -275,40 +226,30 @@ describe('SgidClient', () => {
       expect(actual.searchParams.get('redirect_uri')).toBe(mockRedirectUri)
       expect(actual.searchParams.get('nonce')).toBe(nonce)
       expect(actual.searchParams.get('state')).toBe(mockState)
-      expect(actual.searchParams.get('code_challenge')).toBe(mockCodeChallenge)
-      expect(actual.searchParams.get('code_challenge_method')).toBe(
-        DEFAULT_SGID_CODE_CHALLENGE_METHOD,
-      )
-      // Client ID, scope, response_type, redirect_uri, nonce, state, code_challenge, code_challenge_method
-      expect(actualNumSearchParams).toBe(8)
+      // Client ID, scope, response_type, redirect_uri, nonce, state
+      expect(actualNumSearchParams).toBe(6)
     })
 
     it('should throw when no redirectUri is provided', () => {
       const mockState = 'mockState'
-      const mockCodeChallenge = 'mockCodeChallenge'
 
       const noRedirectUriClient = new SgidClient({
         clientId: MOCK_CLIENT_ID,
         clientSecret: MOCK_CLIENT_SECRET,
         privateKey: MOCK_CLIENT_PRIVATE_KEY,
         hostname: MOCK_HOSTNAME,
+        apiVersion: MOCK_API_VERSION,
       })
 
-      expect(() =>
-        noRedirectUriClient.authorizationUrl({
-          state: mockState,
-          codeChallenge: mockCodeChallenge,
-        }),
-      ).toThrow('No redirect URI registered with this client')
+      expect(() => noRedirectUriClient.authorizationUrl(mockState)).toThrow(
+        'No redirect URI registered with this client',
+      )
     })
   })
 
   describe('callback', () => {
     it('should call token endpoint and return sub and accessToken', async () => {
-      const { sub, accessToken } = await client.callback({
-        code: MOCK_AUTH_CODE,
-        codeVerifier: MOCK_CODE_VERIFIER,
-      })
+      const { sub, accessToken } = await client.callback(MOCK_AUTH_CODE)
 
       expect(sub).toBe(MOCK_SUB)
       expect(accessToken).toBe(MOCK_ACCESS_TOKEN)
@@ -317,31 +258,23 @@ describe('SgidClient', () => {
     it('should throw when no access token is returned', async () => {
       server.use(tokenHandlerNoToken)
 
-      await expect(
-        client.callback({
-          code: MOCK_AUTH_CODE,
-          codeVerifier: MOCK_CODE_VERIFIER,
-        }),
-      ).rejects.toThrow('Authorization server did not return an access token')
+      await expect(client.callback(MOCK_AUTH_CODE)).rejects.toThrow(
+        'Authorization server did not return an access token',
+      )
     })
 
     it('should throw when sub is empty', async () => {
       server.use(tokenHandlerNoSub)
 
-      await expect(
-        client.callback({
-          code: MOCK_AUTH_CODE,
-          codeVerifier: MOCK_CODE_VERIFIER,
-        }),
-      ).rejects.toThrow('Authorization server did not return the sub claim')
+      await expect(client.callback(MOCK_AUTH_CODE)).rejects.toThrow(
+        'Authorization server did not return the sub claim',
+      )
     })
   })
 
   describe('userinfo', () => {
     it('should call userinfo endpoint and return sub and data', async () => {
-      const { sub, data } = await client.userinfo({
-        accessToken: MOCK_ACCESS_TOKEN,
-      })
+      const { sub, data } = await client.userinfo(MOCK_ACCESS_TOKEN)
 
       expect(sub).toBe(MOCK_SUB)
       expect(data).toEqual(MOCK_USERINFO_PLAINTEXT)
@@ -350,9 +283,7 @@ describe('SgidClient', () => {
     it('should return empty data object when no key is returned', async () => {
       server.use(userInfoHandlerNoKey)
 
-      const { sub, data } = await client.userinfo({
-        accessToken: MOCK_ACCESS_TOKEN,
-      })
+      const { sub, data } = await client.userinfo(MOCK_ACCESS_TOKEN)
 
       expect(sub).toBe(MOCK_SUB)
       expect(data).toEqual({})
@@ -361,9 +292,7 @@ describe('SgidClient', () => {
     it('should return empty data object when no data is returned', async () => {
       server.use(userInfoHandlerNoData)
 
-      const { sub, data } = await client.userinfo({
-        accessToken: MOCK_ACCESS_TOKEN,
-      })
+      const { sub, data } = await client.userinfo(MOCK_ACCESS_TOKEN)
 
       expect(sub).toBe(MOCK_SUB)
       expect(data).toEqual({})
@@ -376,107 +305,28 @@ describe('SgidClient', () => {
         privateKey: 'malformed',
         redirectUri: MOCK_REDIRECT_URI,
         hostname: MOCK_HOSTNAME,
+        apiVersion: MOCK_API_VERSION,
       })
 
       await expect(
-        invalidPrivateKeyClient.userinfo({
-          accessToken: MOCK_ACCESS_TOKEN,
-        }),
+        invalidPrivateKeyClient.userinfo(MOCK_ACCESS_TOKEN),
       ).rejects.toThrow('Failed to import private key')
     })
 
     it('should throw when encrypted key is malformed', async () => {
       server.use(userInfoHandlerMalformedKey)
 
-      await expect(
-        client.userinfo({
-          accessToken: MOCK_ACCESS_TOKEN,
-        }),
-      ).rejects.toThrow('Unable to decrypt or import payload key')
+      await expect(client.userinfo(MOCK_ACCESS_TOKEN)).rejects.toThrow(
+        'Unable to decrypt or import payload key',
+      )
     })
 
     it('should throw when encrypted data is malformed', async () => {
       server.use(userInfoHandlerMalformedData)
 
-      await expect(
-        client.userinfo({
-          accessToken: MOCK_ACCESS_TOKEN,
-        }),
-      ).rejects.toThrow('Unable to decrypt payload')
-    })
-  })
-
-  describe('generatePkcePair', () => {
-    it('should generate a PKCE pair when no length is provided', () => {
-      const { codeChallenge, codeVerifier } = generatePkcePair()
-
-      expect(codeVerifier.length).toBe(43)
-      expect(codeChallenge.length).toBe(43)
-
-      expect(codeChallenge).toMatch(codeVerifierAndChallengePattern)
-      expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
-      expect(generateCodeChallenge(codeVerifier)).toBe(codeChallenge)
-    })
-
-    it('should generate a PKCE pair of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
-      for (let length = 43; length <= 128; length++) {
-        const { codeChallenge, codeVerifier } = generatePkcePair(length)
-
-        // Length is only for the code verifier
-        expect(codeVerifier.length).toBe(length)
-
-        expect(codeChallenge).toMatch(codeVerifierAndChallengePattern)
-        expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
-        expect(generateCodeChallenge(codeVerifier)).toBe(codeChallenge)
-      }
-    })
-
-    it('should throw an error when a length < 43 or length > 128 is provided', () => {
-      for (const length of [-1, 0, 42, 129, 138, 999]) {
-        expect(() => generatePkcePair(length)).toThrowError(
-          'generatePkcePair should receive a minimum length of 43 and a maximum length of 128',
-        )
-      }
-    })
-  })
-
-  describe('generateCodeVerifier', () => {
-    it('should generate a code verifier of length 43 when no length is provided', () => {
-      const codeVerifier = generateCodeVerifier()
-
-      expect(codeVerifier.length).toBe(43)
-      expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
-    })
-
-    it('should generate a code verifier of specified length when length between 43 (inclusive) and 128 (inclusive) is provided', () => {
-      for (let length = 43; length <= 128; length++) {
-        const codeVerifier = generateCodeVerifier(length)
-        expect(codeVerifier.length).toBe(length)
-        expect(codeVerifier).toMatch(codeVerifierAndChallengePattern)
-      }
-    })
-
-    it('should throw an error when a length < 43 or length > 128 is provided', () => {
-      for (const length of [-1, 0, 42, 129, 138, 999]) {
-        expect(() => generateCodeVerifier(length)).toThrowError(
-          `Code verifier should have a minimum length of 43 and a maximum length of 128`,
-        )
-      }
-    })
-  })
-
-  describe('generateCodeChallenge', () => {
-    it('should match the specified pattern', () => {
-      expect(generateCodeChallenge(MOCK_CODE_VERIFIER)).toMatch(
-        codeVerifierAndChallengePattern,
+      await expect(client.userinfo(MOCK_ACCESS_TOKEN)).rejects.toThrow(
+        'Unable to decrypt payload',
       )
-    })
-
-    it('should be deterministic (return the same code challenge given the same code verifier)', () => {
-      const firstCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
-      const secondCodeChallenge = generateCodeChallenge(MOCK_CODE_VERIFIER)
-
-      expect(firstCodeChallenge).toBe(secondCodeChallenge)
     })
   })
 })
